requirements.txt
llama-index>=0.10.30
llama-index-llms-openai>=0.1.20
llama-index-embeddings-openai>=0.1.10
llama-index-tools-sql>=0.1.5
sqlalchemy>=2.0.0
python-dotenv>=0.19.0
flask>=2.3.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.0.0
line-bot-sdk>=3.5.0
requests>=2.31.0